refactor(modules/music_player): remove `AutocompleteService` from application layer
